requirements.txt
requests>=2.31.0
pandas>=2.1.0
openpyxl>=3.1.2
PyYAML>=6.0.1
python-dateutil>=2.8.2
openai>=1.40.0
pypdf>=4.0.0
